Criado pacote caixa_eletronico e combinadores [Modificado] -Extraido CombinadorRecursivo de CaixaEletronico.java
