+++ src/Main.java
@@ -1,0 +1,10 @@
+import caixa_eletronico.*;
+import ui.*;
+
+public class Main {
+    public static void main(String[] args) {
+        ICaixaEletronico cx = new CaixaEletronico();
+        //GUI janela = new GUI(cx);
+        //janela.show();
+    }
+}
